analytics/requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
python-dotenv
    A